Changed smart pot variable types.
- Changed the types of the smart pot variable types to only use the amount of space that is required

+++ src/EEPROM_promter.h
@@ -2,9 +2,10 @@
 #define EEPROM_H
 
 #include <avr/io.h>
+#include <stdint.h>
 
 // Function to write data to EEPROM /
-void EEPROM_write(unsigned int uiAddress, unsigned char ucData);
+void EEPROM_write(uint8_t uiAddress, unsigned char ucData);
 
 // Function to read data from EEPROM /
 unsigned char EEPROM_read(unsigned int uiAddress);
